+++ libraries/transformer/src/main/java/androidx/media3/transformer/MuxerWrapper.java
@@ -48,7 +48,6 @@ import androidx.media3.common.util.Util;
 import androidx.media3.container.NalUnitUtil;
 import androidx.media3.effect.DebugTraceUtil;
 import androidx.media3.muxer.Muxer;
-import androidx.media3.muxer.Muxer.TrackToken;
 import androidx.media3.muxer.MuxerException;
 import com.google.common.collect.ImmutableList;
 import java.io.File;
@@ -580,7 +579,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
         data.remaining(),
         presentationTimeUs,
         TransformerUtil.getMediaCodecFlags(isKeyFrame ? C.BUFFER_FLAG_KEY_FRAME : 0));
-    muxer.writeSampleData(trackInfo.trackToken, data, bufferInfo);
+    muxer.writeSampleData(trackInfo.trackId, data, bufferInfo);
 
     DebugTraceUtil.logEvent(
         COMPONENT_MUXER,
@@ -749,16 +748,16 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
   private static final class TrackInfo {
     public final Format format;
-    public final TrackToken trackToken;
+    public final int trackId;
 
     public long startTimeUs;
     public long bytesWritten;
     public int sampleCount;
     public long timeUs;
 
-    public TrackInfo(Format format, TrackToken trackToken) {
+    public TrackInfo(Format format, int trackId) {
       this.format = format;
-      this.trackToken = trackToken;
+      this.trackId = trackId;
     }
 
     /**